requirements.txt
sounddevice==0.4.6
soundfile==0.12.1
PyQt6==6.4.2
numpy==1.26.0
openai-whisper>=20240930
deep-translator==1.11.4
